tests/wpt/web-platform-tests/origin-isolation/parent-yes-child-yes-subdomain.sub.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Parent is isolated, child is not isolated, child is a subdomain of the parent</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<div id="log"></div>

<script type="module">
import { insertIframe, sendWasmModule, setBothDocumentDomains } from "./resources/helpers.mjs";

let frameWindow;
promise_setup(async () => {
  frameWindow = await insertIframe("{{hosts[][www]}}", "?1");
});

// Both request isolation, so the parent ends up in one origin-keyed agent
// cluster (the base domain's origin), and the child ends up in a different
// origin-keyed agent cluster (the www subdomain's origin).

promise_test(async () => {
  const whatHappened = await sendWasmModule(frameWindow);

  assert_equals(whatHappened, "messageerror");
}, "messageerror event must occur");

promise_test(async () => {
  await setBothDocumentDomains(frameWindow);

  assert_throws_dom("SecurityError", DOMException, () => {
    frameWindow.document;
  });
}, "setting document.domain must not give sync access");
</script>
